+++ s-favorite-cities/package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@chakra-ui/icons": "^2.2.4",
     "@chakra-ui/react": "^3.1.1",
     "@emotion/react": "^11.13.3",
     "next": "15.0.3",
