+++ s-favorite-cities/pages/_app.js
@@ -1,5 +1,8 @@
 import { ChakraProvider, defaultSystem } from "@chakra-ui/react";
-import Navbar from "../components/Navbar"; // Adjust the path if needed
+// import Navbar from "../components/Navbar";
+// import "../styles/globals.css";
+// import Navbar from "../components/NavbarChakraTep";
+import Navbar from "../components/Navbarchakraui";
 
 export default function App({ Component, pageProps: { ...pageProps } }) {
   return (
