+++ s-favorite-cities/pages/favorites.jsx
@@ -1,11 +1,14 @@
 import React from "react";
+import { Container, Heading } from "@chakra-ui/react";
 
 const Favorites = () => {
   return (
-    <div>
-      <h1>Your Favorite Cities</h1>
+    <Container maxW="2xl">
+      <Heading my="2rem" p="0.7rem" ml="2.2rem">
+        Your favorite cities ❤:
+      </Heading>
       {/* Display the list of favorite cities */}
-    </div>
+    </Container>
   );
 };
