+++ s-favorite-cities/pages/index.jsx
@@ -1,11 +1,17 @@
 import React from "react";
+import { Container, Heading, Text } from "@chakra-ui/react";
 
 const Homepage = () => {
   return (
-    <div className="main">
-      <h1>Welcome to the Favorite Cities App!</h1>
-      <p>Select a destination or search for a new city.</p>
-    </div>
+    <Container maxW="2xl">
+      <Heading my="2rem" p="0.7rem" ml="2.2rem">
+        Welcome to the Favorite Cities App! 🌆
+      </Heading>
+      <Text marginLeft="2rem">
+        A nice app developed for you, to make a plan for each destination you
+        have or to search for a new city.
+      </Text>
+    </Container>
   );
 };
